refactor: move CallProgramHandler tests in the new suite CallProgramHandlerTest
